Exercise 85: Reformatory
The reformatory class has three methods:
 - weight(Person person) method:
The method gets a person object as a parameter. The method is supposed to return the weight of the parameter, so the method should call a suitable method of person, get the return value and then return it to the caller.
- feed(Person person):
It increases the weight of its parameter by one.
- totalWeightsMeasured()
It returns the total number of times a weight has been measured.