Exercise 101: The library information system
It's a simple information system prototype for a library. 
The prototype has functionality for searching books by the title, publisher or publishing year.

Book class: 
- Objects of the class Book represent the information of a single book. 
Instance variables: 
   - title (String);
   - publisher (String);
   - year (int);
Constructor:
   - Book(String title, String publisher, int year); 
Methods: 
   - String getTitle(), 
   - String getPublisher(), 
   - int getYear() 
   - String toString().


Library class:  
- Object of the class Library holds a set of books and provides various ways to search for the books within the library.
Instance variable: 
   - catalog  (ArrayList<Book>)
Constructor:
  - Library()
Methods: 
  - void addBook(Book newBook)
  - public void printBooks()
  - ArrayList<Book> searchByTitle(String title)
  - ArrayList<Book> searchByPublisher(String publisher)
  - ArrayList<Book> searchByYear(int year)
The methods return the list of books that match the given title, publisher or year.


StringUtils class: 
Method 
   - public static boolean included(String word, String searched)
It checks if the string searched is contained within the string word. 
The method is case insensitive and doesn't care about trailing and ending white spaces in the string searched. 
If either of the strings is null, the method should return false.