Exercise 102: Grade distribution
The input of the program is a set of exam scores of a course. 
Each score is an integer. 
When -1 is entered, the program stops asking for further input.

After the scores have been read, the program prints the grade distribution and acceptance percentage of the course in the following form:
Grade distribution:
5: **
4:
3: ***
2: *
1: *
0: *

Grade distribution is formed as follows:
Each exam score is mapped to a grade: 
0-29 => 0
30-34 => 1
35-39 => 2
40-44 => 3
45-49 => 4
50-60 => 5

If the score is not within the range 0-60 it is not taken into account.
The number of grades are printed as stars.
If there are no scores that correspond to a particular grade, the printed line  doesn't have stars. 

At the end the program must print the Acceptance percentage.
All the grades besides zeros are accepted. 
Acceptance percentage is calculated with the formula:  100*accepted/allScores.