Exercise 105: Guessing game
A program which guesses the number the player is thinking about. 
The program assumes that the number is between lowerLimit...upperLimit. 
The start of the game provides these limits to the method as parameters that makes the game happen.
The program asks the player questions in the format "Is your number greater than X?" and deduce the correct answer from the answers the player gives.

The program keeps track of the search area with the help of the variables lowerLimit and upperLimit. 
The program always asks if the player's number is greater than the average of these two numbers, and based on the answers the search area gets halved each time. 
In the end the lowerLimit and upperLimit are the same and the number the user is thinking of has been revealed.

In the following example the user chooses the number 44:

Think of a number between 1...100.
I promise you that I can guess the number you are thinking of with 7 questions.

Next I'll present you a series of questions. Answer them honestly.

Is your number greater than 50? (y/n)
n
Is your number greater than 25? (y/n)
y
Is your number greater than 38? (y/n)
y
Is your number greater than 44? (y/n)
n
Is your number greater than 41? (y/n)
y
Is your number greater than 43? (y/n)
y
The number you're thinking of is 44.
  

- method isGreaterThan(int value) presents the user with a question:
"Is your number greater than given value? (y/n)"

- a method average(int firstNumber, int secondNumber) calculates the average of the given values.