Exercise 3: Calculator
It is a simple calculator.
The main program just creates the object that implements the actual application logic and then starts it. 

A class Reader encapsulates a Scanner-object. It has the following elements:
- public String readString() method;
- public int readInteger() method;

A class Calculator takes care of the application logic of the program. 
It has
- a method public void start().
- statistics() method prints the amount of operations done with the Calculator-object

The calculator has the operations:
- sum, 
- difference, 
- product

- end

First, the program asks user for the command and two numbers to work with.
If an invalid command is given to the program (something other than sum, difference, product or end), the calculator will not react to the command in any way, but instead continues by asking the next command. 
Statistics is not to count an invalid command as a completed calculation.