Exercise 4: Thing, Suitcase, and Container
Class Thing:
The class objects can represent different kinds of things. 
The information to store are the thing's name and weight (kg).
The class contains the following functionality:
- a construsctor, which is given the thing's name and weight as parameter
- public String getName() method, which returns the thing's name
- public int getWeight() method, which returns the thing's weight
- public String toString() method, which returns a string in the form "name (weight kg)"

Class Suitcase: 
Suitcase has things and a maximum weight limit, which defines the greatest total allowed weight of the things contained within the Suitcase object.
The class has the following methods:
- a constructor, which is given a maximum weight limit
- public void addThing(Thing thing), which adds the thing in the parameter to your suitcase. The method does not return any value.
- public String toString(), which returns a string in the form "x things (y kg)"
- printThings, which prints out all the things inside the suitcase
- totalWeight, which returns the total weight of the things in your suitcase
- heaviestThing, which returns the thing which weighs the most (if there are more than one thing with the same weight, the method can return either one, if the suitcase is empty, the method returns null).
The things are saved into an ArrayList object:
The class Suitcase makes sure the thing's weight does not cause the total weight to exceed the maximum weight limit. 
The method addThing does not add a new thing if the total weight happens to exceed the maximum weight limit.
toString output looks like this (in case of 0 things and 1 thing the spelling is appropriate):
E.g:
empty (0 kg)
1 thing (2 kg)
2 things (3 kg)
2 things (3 kg)
            
Class Container:
The class Container has the following methods:
- a constructor which is given the maximum weight limit;
- public void addSuitcase(Suitcase suitcase), which adds the suitcase as a parameter to the container
- public String toString() which returns a string in the form "x suitcases (y kg)"
-  public void printThings() method prints out all the things inside the container's suitcases.
The suitcases are stored in the ArrayList.
The class Container has to make sure the thing's total weight does not overcome the maximum weight limitation. 
The method addSuitcase should not add a new suitcase if the total weight happens to exceed the maximum weight limit.

Main class:  
In the Main class there is the method public static void addSuitcasesFullOfBricks(Container container), which adds 100 suitcases into the container it receives as parameter; there is one brick in each suitcase. 
The bricks weight will then increase by one each time until the weight becomes 100 kg.