Exercise 7: Dictionary
Exercise: 
- Create a class called Dictionary. The class has the following methods:
  - public String translate(String word), returning the translation of its parameter. If the word is unknown, it returns null.
  - public void add(String word, String translation), adding a new translation to the dictionary

Implement the class Dictionary so that it contained only one object variable, a HashMap data structure.
  - add the method public int amountOfWords(), which returns the amount of words in the dictionary.
  - add the method public ArrayList<String> translationList() to your dictionary, returning strings which stand for a content list of your dictionary in the form key = value.

Create the class TextUserInterface, with the following methods:
  - the constructor public TextUserInterface(Scanner reader, Dictionary dictionary)
  - public void start(), which starts the interface.
The text user interface stores into two object variables the reader and dictionary it has received as constructor parameters. 
You don't need other object variables. 
The user input must be read using the reader object received as constructor parameter! 
The translations also have to be stored into the dictionary object received as constructor parameter. 
The text user interface must not create new objects itself!

At the beginning, in the text user interface must only have the command quit, to quit the text user interface. 
If the user inputs something else, we print "Unknown statement".

Add the methods add() and translate() to your text user interface. 
The command add asks for a word pair from the user and adds them to the dictionary. 
The command translate asks a word from the user and it prints the translation.