Exercise 8: Airport
Exercise: 
Create an application to manage an airport, with its airplanes and flights. 
As far as the planes are concerned, it's always known their ID and capacity. 
As for the flights, it's known the plane used for the flight, the departure airport code (for instance HEL) and the destination airport code (for instance BAL).

There can be various different flights and planes. 
The same plane can also go on various different flights (various different routes). 
The application must offer two different panels. 
First, the airport worker inputs the flight and plane information to the system in the airport panel.

When the user exits the airport panel, the user then proceeds to use the flight service. 
The flight service has three actions: printing planes, printing flights, and printing airplane information. 
In addition to this, the user can exit the application by choosing x. 
If the user inputs an invalid command, the command is asked again.