Exercise 12: Online Shop
The class Storehouse has the following methods:
  - public void addProduct(String product, int price, int stock), adding to the storehouse a product whose price and number of stocks are the parameter values
  - public int price(String product) returns the price of the parameter product; if the product is not available in the storehouse, the method returns -99
Inside the storehouse, the prices and the stocks of the products are stored into a Map<String, Integer> variable.
  - public int stock(String product) returns the stock of the parameter product.
  - public boolean take(String product) decreases the stock of the parameter product by one, and it returns true if the object was available in the storehouse. 
If the product was not in the storehouse, the method returns false, the product stock cannot go below zero.
  - public Set<String> products() returns a name set of the products contained in the storehouse
            
A Purchase is a specific number of a specific product. 
For instance, we can put into our shopping basket either a purchase corresponding to one bread, or a purchase corresponding to 24 coffees.
The class Purchase has the following functionality:
  - a constructor public Purchase(String product, int amount, int unitPrice), which creates a purchase corresponding the parameter product. 
The product unit amount of purchase is clarified by the parameter amount, and the third parameter is the unit price
public int price(), which returns the purchase price. 
This is obtained by raising the unit amount by the unit price
  - public void increaseAmount() increases by one the purchase unit amount
  - public String toString() returns the purchase in a string form like the following

The ShoppingBasket stores its products as Purchase objects. 
The shopping basket has an object products variable whose type is either Map<String, Purchase>. 
The ShoppingBasket class has the following functionality: 
  - a constructor without parameter 
  - public void add(String product, int price) adds a purchase to the shopping basket; the purchase is defined by the parameter product, and its price is the second parameter.
  - public int price() returns the shopping basket total price
  - public void print( prints out the Purchase objects which are contained by the basket. 
The printing order is not important. 
If the ShoppingBasket already contains the product which was add, a new Purchase object is not created. Instead, the existing Purchase object corresponding to the existing product is updated by calling its method increaseAmount().

The class Shop to manages the purchases and stock corresponding to the products that customer adds to the basket.