Exercise 15: Sorting Cards
The Cards class has a value and a suit variables. 
Card values are 2, 3, ..., 10, J, Q, K and A, and the suits are Spades, Hearts, Diamonds and Clubs. 
Value and suit are however shown as integers in the objects. 
Cards have also got a toString method, which is used to print the card value and suit in a "friendly way".
Four constants -- that is public static final variables -- are defined in the class, so that the user didn't need to handle card's suits as numbers.

Cards class implements the compareTo method so that cards would be sorted in ascending order according to their value. 
If the value of two classes have got the same values, we compare them against their suit in ascending order: spades first, diamonds second, hearts third, and clubs last.

The class Hand represents the player hand set of cards. 
Hand class has the following methods: 
  - public void add(Card card) adds a card to the hand
  - public void print() prints the cards in the hand
  - public void sort() sorts the cards in the hand
The class SortAgainstSuitAndValue implements the Comparator interface and sorts cards by value and value

  - method public void sortAgainstSuit() in the class Hand sorts cards according to the comparator SortAgainstSuitAndValue