Exercise 20: Sensors and Temperature Measurement
All the code is placed in the package application.

There is a Sensor interface.
      
A class ConstantSensor implements the interface Sensor.
The constant sensor is online all the time. 
The methods on() and off() do not do anything. 
The constant sensor has a constructor with an int parameter. 
The measure method call returns the number received as constructor parameter.

A class Thermometer implements the interface Sensor.
At first, the thermometer is on. 
When the measure method is called, if the thermometer is on it returns a random number between -30 and 30. 
If the thermometer is off, it throws an IllegalStateException.

A class AverageSensor implements the interface Sensor.
An average sensor contains many sensors. 
In addition to the methods defined by the interface Sensor, the class has the method public void addSensor(Sensor additional) which adds a new sensor to the AverageSensor.
The average sensor is on when all its sensors are on. 
When the average sensor is switched on, all its sensors have to be switched on if they were not on already. 
When the average sensor is closed, at least one of its sensors has to be switched off. 
It's also possible that all its sensors are switched off.
The measure method of AverageSensor returns the average of the readings of all its sensors. 
If the measure method is called when the average sensor is off, or if the average sensor was not added any sensor, the method throws an IllegalStateException.
The method public List<Integer> readings() in AverageSensor returns a list of the reading results of all the measurements executed through your AverageSensor.