Exercise 27: The Finnish Ringing Center
The Finnish Ringing Center stores the information about the birds who were watched in a specific year in Bird objects:

The idea is implementing the functionality for the ringing center to track of the places where rung birds were observed and how many times they were. 
However, observation places and times are not stored in Bird objects, but in a separate HashMap, whose keys are Bird objects. 

Some birds have more than one English name (for instance, the Rose Starling is also known as Rose-Coloured Starling or Rose-Coloured Pastor); however, the Latin name is always unique. 
Two Bird objects are understood as the same bird if their Latin name and observation year are the same.

The Ringing Center has two methods: 
  - public void observe(Bird bird, String place), which records the observation and its place among the bird information; 
  - public void observations(Bird bird), which prints all the observations of the parameter bird.