Exercise 28: Groups
In this mini-program organisms and groups of organisms move around. 
The position of the organisms is reported by using a bidimensional coordinate system. 
Each position is defined by two numbers, the x and y coordinates. The x coordinate tells us how far from the "point zero" the position is horizontally, whereas the y coordinate tells u how far the position is vertically.

The interface Movable represents things that can be moved from one place to another. 
The interface contains the method void move(int dx, int dy). 
The parameter dx tells us how much the object moves on the x axis and dy tells us about the movement on the y axis.

The class Organism is in the package movable.
Organism implement the interface Movable. 
Organisms have to know their own position (x and y coordinates). 
The API of Organism is the following:
  - the constructor public Organism(int x, int y); it receives the x and y initial coordinates of the object
  - public String toString(); it creates and returns a string which represents the object. The form should be the following "x: 3; y: 6". 
  - public void move(int dx, int dy); it moves the object as much as it is specified by the arguments. The variable dx contains the x coordinate of the movement, whereas dy contains the y coordinate of the movement.

The class Group is also in the package movable. 
Group implements the interface Movable. 
The Group is made of various different objects which implement the interface Movable, and they are stored into a list construction
The class Group has the following API:
  - public String toString(); it returns a string which describes the position of the group organisms, each organism is printed in its own line.
  - public void addToGroup(Movable movable); it adds a new objects which implements the interface Movable to the group.
  - public void move(int dx, int dy); it moves a group as much as it is defined by the arguments. Note that you will have to move each group organism.