Exercise 33: Dungeon
Game description: 
In the game, the player is in a dungeon full of vampires. 
The player has to destroy the vampires before his lamp runs out of battery and the vampires can suck his blood in the darkness. 
The player can see the vampires with a blinking of their lamp, after which they have to move blind before the following blinking. 
With one move, the player can walk as many steps as they want.

The game situation, i.e. the dungeon, the player and the vampires are shown in text form. 
Program print output tells how many moves the player has left (that is to say, how much battery the lamp has). 
Also, it shows player and vampire positions on a game map. As on the example below: 

P: 0, 0
V: 7, 0
V: 3, 0
V: 5, 4
V: 2, 4
P--V---V---
-----------
-----------
-----------
--V--V-----
Moves left: 20
Your step(s): dd


The example above shows the lamp has enough battery for 20 blinkings. 
The player P is located at 0,0. Note that the coordinates are calculated starting from the high left corner of the game board. I
The user can move by giving a sequence of commands and pressing Enter. The commands are:
w   - go up
s    - go down
a    - go left
d    -  go right
When the user commands are executed (the user can give many commands at once), a new game situation is drawn. 
If the lamp charge reaches 0, the game ends and the text YOU LOST! GAME OVER is printed on the board.

The vampires move randomly in the game, and they take one step for each step the player takes. 
If the player and a vampire run into each other (even momentarily) the vampire is destroyed. 
If a vampire tries to step outside the board, or into a place already occupied by another vampire, the move is not executed. 
When all the vampires are destroyed, the game ends and it prints YOU WON.

The player always starts the game in the position 0,0.
Player and vampires can not move out of the dungeon and two vampires cannot step into the same place.