Exercise 47: Flexible Filtering Criteria
This program filters the lines of the books found on the Project Gutenberg pages. 
In the provided example Dostoevsky's Crime and Punishment is analyzed. 
There are various different filtering criteria, and it is possible to filter according to different criteria combinations. 
Also, the program structure allows to add new criteria later on.