Exercise 48: String builder
public static String build(int[] t) modifies parameter array so that it would look like:

{
1, 2, 3, 4,
5, 6, 7, 8,
9, 10
}
  
The curly brackets are on their own lines. 
There are up to four values in each line of the table, and the first value requires an initial space. 
Before each number, and after the comma there is exactly one space.